fix: Se corrige consulta de información juego lectura y se añade comportamiento de erro en frontend

+++ backend/app/validators/strategy-read.validator.js
@@ -24,8 +24,7 @@ const validateGroup = [
     .trim()
     .bail()
     .isIn(['AGES', 'TonicaAtona', 'Tritonica'])
-    .withMessage('Select valid theme [AGES,TonicaAtona,Tritonica]')
-    .toInt(),
+    .withMessage('Select valid theme [AGES,TonicaAtona,Tritonica]'),
 
   (req, res, next) => {
     validateResult(req, res, next)
